requirements.txt
python-bitcoinlib
Pi7SegPy                                              
~                            
